+++ app/javascript/vue/tasks/nomenclature/match/components/Table/TableMatched.vue
@@ -16,7 +16,15 @@
         <td>
           <span v-html="taxon.object_tag" /> {{ taxon.cached_author_year }}
         </td>
-        <td v-html="taxon.original_combination" />
+        <td>
+          <template v-if="taxon.cached_is_valid">
+            <span v-html="taxon.object_tag" /> {{ taxon.cached_author_year }}
+          </template>
+          <template v-if="validNames[taxon.cached_valid_taxon_name_id]">
+            <span v-html="validNames[taxon.cached_valid_taxon_name_id].object_tag" />
+            {{ validNames[taxon.cached_valid_taxon_name_id].cached_author_year }}
+          </template>
+        </td>
       </tr>
     </tbody>
   </table>
@@ -27,6 +35,12 @@ defineProps({
   list: {
     type: Array,
     required: true
+  },
+
+  validNames: {
+    type: Object,
+    default: () => ({})
   }
 })
+
 </script>